COMP: Fix in sign-compare warning in vtkITKTransformConverter
This commit fixes the following warning:

/home/kitware/Dashboards/Nightly/Slicer-0/Libs/MRML/Core/vtkITKTransformConverter.h: In static member function 'static bool vtkITKTransformConverter::SetITKImageFromVTKOrientedGridTransform(vtkObject*, itk::Image<itk::Vector<double, 3u>, 3u>::Pointer&, vtkOrientedGridTransform*)':
/home/kitware/Dashboards/Nightly/Slicer-0/Libs/MRML/Core/vtkITKTransformConverter.h:1011:53: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=26803
git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-8@26803 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkITKTransformConverter.h
@@ -1008,7 +1008,7 @@ bool vtkITKTransformConverter::SetITKImageFromVTKOrientedGridTransform(vtkObject
     vtkErrorWithObjectMacro(loggerObject, "Cannot save grid transform: the input vtkOrientedGridTransform does not contain a valid displacement grid");
     return false;
     }
-  if (gridImage_Ras->GetNumberOfScalarComponents()!=VTKDimension)
+  if (gridImage_Ras->GetNumberOfScalarComponents() != static_cast<int>(VTKDimension))
     {
     vtkErrorWithObjectMacro(loggerObject, "Cannot save grid transform: the input vtkOrientedGridTransform expected to contain "
       << VTKDimension << " components but it actually contains " << gridImage_Ras->GetNumberOfScalarComponents() );
